+++ sciris/sc_version.py
@@ -4,6 +4,6 @@ Version and license information.
 
 __all__ = ['__version__', '__versiondate__', '__license__']
 
-__version__      = '3.1.3'
-__versiondate__  = '2024-02-07'
+__version__      = '3.1.4'
+__versiondate__  = '2024-03-11'
 __license__      = f'Sciris {__version__} ({__versiondate__}) – © 2014-2024 by the Sciris Development Team'